requirements.txt
fastapi == 0.115.0
pydantic == 2.9.2
uvicorn == 0.31.0
python-dotenv == 1.0.1
langchain-core == 0.3.15
langchain-ollama == 0.2.0
langchain-openai == 0.2.6
openai == 1.55.1